# backend/backend/openapi.yaml
openapi: 3.0.3
info:
  title: PetClinic API
  version: 1.0.0
  description: |
    REST API for the Spring PetClinic schema (TypeScript / Express / Sequelize port).

    Interactive docs: `/api/docs`. Raw spec: `/api/docs.json`.

    **Authentication.** Every `/api/*` endpoint requires a Bearer access token
    obtained from `POST /auth/login` or `POST /auth/signup`. Unauthenticated
    requests get `401`. The `/auth/*` endpoints, `/health`, and the docs are
    public.
servers:
  - url: /
    description: Same origin as the running server.
security:
  - bearerAuth: []
...
paths:
...
  /api/owners:
    get:
      tags: [Owners]
      summary: List owners
      parameters:
        - in: query
          name: lastName
          required: false
          schema: { type: string }
          description: Prefix match on `last_name` (case-sensitive in SQLite). Owners are returned with their pets, each pet's type, and each pet's visits.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Owner' }
    post:
      tags: [Owners]
      summary: Create owner
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/OwnerInput' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Owner' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/owners/{ownerId}:
    parameters:
      - $ref: '#/components/parameters/OwnerId'
    get:
      tags: [Owners]
      summary: Get owner by id (includes pets, pet types, and visits)
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Owner' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [Owners]
      summary: Update owner
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/OwnerInput' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Owner' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
    delete:
      tags: [Owners]
      summary: Delete owner
      responses:
        '204': { description: No Content }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/owners/{ownerId}/pets:
    parameters:
      - $ref: '#/components/parameters/OwnerId'
    post:
      tags: [Owners, Pets]
      summary: Create a pet for this owner
      description: '`owner_id` is taken from the URL; do not send it in the body.'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PetInputForOwner' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Pet' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/owners/{ownerId}/pets/{petId}:
    parameters:
      - $ref: '#/components/parameters/OwnerId'
      - $ref: '#/components/parameters/PetId'
    get:
      tags: [Owners, Pets]
      summary: Get a pet belonging to this owner
      description: 404 if the pet exists but belongs to a different owner.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Pet' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [Owners, Pets]
      summary: Update a pet belonging to this owner
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PetInputForOwner' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Pet' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/owners/{ownerId}/pets/{petId}/visits:
    parameters:
      - $ref: '#/components/parameters/OwnerId'
      - $ref: '#/components/parameters/PetId'
    get:
      tags: [Owners, Visits]
      summary: List visits for a pet belonging to this owner
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Visit' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
    post:
      tags: [Owners, Visits]
      summary: Add a visit to a pet belonging to this owner
      description: '`pet_id` is taken from the URL; do not send it in the body.'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/VisitInputForPet' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Visit' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
...
  /api/pets/{petId}:
    parameters:
      - $ref: '#/components/parameters/PetId'
    get:
      tags: [Pets]
      summary: Get pet by id (with type, owner, and visits included)
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Pet' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [Pets]
      summary: Update pet
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PetInput' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Pet' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
    delete:
      tags: [Pets]
      summary: Delete pet
      responses:
        '204': { description: No Content }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
...
  /api/pettypes:
    get:
      tags: [PetTypes]
      summary: List pet types (alphabetical)
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/PetType' }
    post:
      tags: [PetTypes]
      summary: Create pet type
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PetTypeInput' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PetType' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/pettypes/{petTypeId}:
    parameters:
      - $ref: '#/components/parameters/PetTypeId'
    get:
      tags: [PetTypes]
      summary: Get pet type by id
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PetType' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [PetTypes]
      summary: Update pet type
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PetTypeInput' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PetType' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
    delete:
      tags: [PetTypes]
      summary: Delete pet type
      responses:
        '204': { description: No Content }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/specialties:
    get:
      tags: [Specialties]
      summary: List specialties (alphabetical)
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Specialty' }
    post:
      tags: [Specialties]
      summary: Create specialty
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SpecialtyInput' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Specialty' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/specialties/{specialtyId}:
    parameters:
      - $ref: '#/components/parameters/SpecialtyId'
    get:
      tags: [Specialties]
      summary: Get specialty by id
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Specialty' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [Specialties]
      summary: Update specialty
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SpecialtyInput' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Specialty' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
    delete:
      tags: [Specialties]
      summary: Delete specialty
      responses:
        '204': { description: No Content }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token issued by `/auth/login`, `/auth/signup`, or `/auth/refresh`.
  parameters:
    OwnerId:
      in: path
      name: ownerId
      required: true
      schema: { type: integer, minimum: 0 }
    PetId:
      in: path
      name: petId
      required: true
      schema: { type: integer, minimum: 0 }
...
    SpecialtyId:
      in: path
      name: specialtyId
      required: true
      schema: { type: integer, minimum: 0 }
    PetTypeId:
      in: path
      name: petTypeId
      required: true
      schema: { type: integer, minimum: 0 }
  responses:
    BadRequest:
      description: Malformed request (bad path param, missing required field, validation failure).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Conflict:
      description: Unique-constraint or foreign-key violation.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          example: Owner with id 99 not found
        details:
          type: array
          items: { type: string }
          description: Per-field validation messages (Sequelize validation / unique errors).
        path:
          type: string
          description: Returned for the global "Route not found" 404.
    Owner:
      type: object
      required: [id]
      properties:
        id: { type: integer, readOnly: true, example: 1 }
        first_name: { type: string, nullable: true, example: George }
        last_name:  { type: string, nullable: true, example: Franklin }
        address:    { type: string, nullable: true, example: 110 W. Liberty St. }
        city:       { type: string, nullable: true, example: Madison }
        telephone:  { type: string, nullable: true, example: '6085551023' }
        pets:
          type: array
          items: { $ref: '#/components/schemas/Pet' }
          description: Present when fetched via the owners endpoints (which include pets, pet types, and visits).
    OwnerInput:
      type: object
      properties:
        first_name: { type: string, nullable: true }
        last_name:  { type: string, nullable: true }
        address:    { type: string, nullable: true }
        city:       { type: string, nullable: true }
        telephone:  { type: string, nullable: true }
    Pet:
      type: object
      required: [id, type_id, owner_id]
      properties:
        id:         { type: integer, readOnly: true, example: 1 }
        name:       { type: string, nullable: true, example: Leo }
        birth_date: { type: string, format: date, nullable: true, example: '2010-09-07' }
        type_id:    { type: integer, example: 1 }
        owner_id:   { type: integer, example: 1 }
        type:
          allOf:
            - $ref: '#/components/schemas/PetType'
          description: Included via Sequelize association.
        owner:
          allOf:
            - $ref: '#/components/schemas/Owner'
          description: Included via Sequelize association.
        visits:
          type: array
          items: { $ref: '#/components/schemas/Visit' }
          description: Included via Sequelize association.
    PetInput:
      type: object
      description: Body for `PUT /api/pets/{petId}`.
      properties:
        name:       { type: string, nullable: true }
        birth_date: { type: string, format: date, nullable: true }
        type_id:    { type: integer }
        owner_id:   { type: integer }
    PetInputForOwner:
      type: object
      description: Body for pet endpoints nested under `/api/owners/{ownerId}`. `owner_id` is taken from the URL.
      required: [type_id]
      properties:
        name:       { type: string, nullable: true }
        birth_date: { type: string, format: date, nullable: true }
        type_id:    { type: integer }
...
    Specialty:
      type: object
      required: [id]
      properties:
        id:   { type: integer, readOnly: true }
        name: { type: string, nullable: true }
    SpecialtyInput:
      type: object
      properties:
        name: { type: string, nullable: true }
    PetType:
      type: object
      required: [id]
      properties:
        id:   { type: integer, readOnly: true }
        name: { type: string, nullable: true }
    PetTypeInput:
      type: object
      properties:
        name: { type: string, nullable: true }
    Visit:
      type: object
      required: [id, pet_id]
      properties:
        id:          { type: integer, readOnly: true }
        pet_id:      { type: integer }
        visit_date:  { type: string, format: date, nullable: true }
        description: { type: string, nullable: true }
        pet:
          allOf:
            - $ref: '#/components/schemas/Pet'
          description: Included via Sequelize association.
...
    VisitInputForPet:
      type: object
      description: Body for visit endpoints nested under a pet. `pet_id` is taken from the URL.
      properties:
        visit_date:  { type: string, format: date, nullable: true }
        description: { type: string, nullable: true }
...